move getAdminToken() into new BaseTest class.
