+++ src/main/java/test/courses/WHAT_153.java
@@ -1,7 +1,6 @@
 package test.courses;
 
 import io.restassured.http.ContentType;
-import io.restassured.response.Response;
 import org.testng.annotations.Test;
 
 import java.util.HashMap;
@@ -10,26 +9,13 @@ import java.util.Map;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.*;
 
-public class WHAT_153 {
-    public String getAdminToken() {
-        Map<String, String> credMap = new HashMap<>();
-
-        credMap.put("email", "admin.@gmail.com");
-        credMap.put("password", "admiN_12");
-        Response response = given().
-                body(credMap).
-                contentType(ContentType.JSON).
-                when().
-                post("https://whatbackend.azurewebsites.net/api/accounts/auth");
-        System.out.println(response.getHeader("authorization"));
-        return response.getHeader("authorization");
-    }
+public class WHAT_153 extends BaseTest{
 
     @Test
     public void createCourse200Admin() {
         Map <String, String> courseMap = new HashMap<>();
-        courseMap.put("name","uniqueTest4");
-        String newCourseID = "43";
+        courseMap.put("name","uniqueTest25");
+        String newCourseID = "46";
 
         given().
                 header("Authorization",getAdminToken()).
@@ -37,10 +23,8 @@ public class WHAT_153 {
                 body(courseMap).
                 when().post("https://whatbackend.azurewebsites.net/api/courses").
                 then().assertThat().statusCode(200).
-                and().
-                body("id",hasToString(newCourseID)).
-                and().
-                body("name",hasToString(courseMap.get("name"))).
-                log().body();
+                and().body("id",hasToString(newCourseID)).
+                and().body("name",hasToString(courseMap.get("name"))).
+                and().log().body();
     }
 }
